+++ Core.ApplicationServices/Model/System/SharedSystemUpdateParameters.cs
@@ -17,6 +17,5 @@ namespace Core.ApplicationServices.Model.System
         public Maybe<IEnumerable<UpdatedExternalReferenceProperties>> ExternalReferences { get; set; } = Maybe<IEnumerable<UpdatedExternalReferenceProperties>>.None;
         public OptionalValueChange<Guid?> BusinessTypeUuid { get; set; } = OptionalValueChange<Guid?>.None;
         public OptionalValueChange<IEnumerable<Guid>> TaskRefUuids { get; set; } = OptionalValueChange<IEnumerable<Guid>>.None;
-        public OptionalValueChange<IEnumerable<LicensingAndCodeModel>> LicensingAndCodeModels { get; set; } = OptionalValueChange<IEnumerable<LicensingAndCodeModel>>.None;
     }
 }
