+++ Core.DomainModel/ItSystem/ItSystem.cs
@@ -310,15 +310,5 @@ namespace Core.DomainModel.ItSystem
             };
         }
 
-        public Maybe<OperationError> SetLicensingAndCodeModels(IEnumerable<LicensingAndCodeModel> newModels) {
-            if (newModels.Any(x => x == LicensingAndCodeModel.Proprietary) && newModels.Any(y => y != LicensingAndCodeModel.Proprietary)) 
-                return new OperationError($"The Proprietary licensing and code model cannot be set on the same system as another model", OperationFailure.Conflict);
-
-            LicensingAndCodeModels = [.. newModels.Distinct()];
-
-            return Maybe<OperationError>.None;
-        }
-
-        public ICollection<LicensingAndCodeModel> LicensingAndCodeModels { get; set; } = new List<LicensingAndCodeModel>();
     }
 }
