+++ Presentation.Web/Models/API/V2/Request/System/Regular/CreateItSystemRequestDTO.cs
@@ -75,9 +75,5 @@ namespace Presentation.Web.Models.API.V2.Request.System.Regular
         /// Determines if the system has been deactivated from being taken into use
         /// </summary>
         public bool Deactivated { get; set; }
-        /// <summary>
-        /// The licensing and code model used for this system. Existing models will be replaced by the input data.
-        /// </summary>
-        public IEnumerable<LicensingAndCodeModelChoice>? LicensingAndCodeModels { get; set; }
     }
 }
