+++ Presentation.Web/Models/API/V2/Request/System/Regular/UpdateItSystemRequestDTO.cs
@@ -69,9 +69,5 @@ namespace Presentation.Web.Models.API.V2.Request.System.Regular
         /// </summary>
         public RegistrationScopeChoice? Scope { get; set; }
         public RecommendedArchiveDutyRequestDTO? RecommendedArchiveDuty { get; set; }
-        /// <summary>
-        /// The licensing and code model used for this system. Existing models will be replaced by the input data.
-        /// </summary>
-        public IEnumerable<LicensingAndCodeModelChoice>? LicensingAndCodeModels { get; set; }
     }
 }
