+++ Presentation.Web/Models/API/V2/Request/SystemUsage/BaseItSystemUsageWriteRequestDTO.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using Presentation.Web.Models.API.V2.Request.Generic.Roles;
+using Presentation.Web.Models.API.V2.Types.System;
 
 namespace Presentation.Web.Models.API.V2.Request.SystemUsage
 {
@@ -22,5 +23,9 @@ namespace Presentation.Web.Models.API.V2.Request.SystemUsage
         /// GDPR-specific registrations
         /// </summary>
         public GDPRWriteRequestDTO? GDPR { get; set; }
+        /// <summary>
+        /// The licensing and code models used for this system usage. Existing models will be replaced by the input data.
+        /// </summary>
+        public IEnumerable<LicensingAndCodeModelChoice>? LicensingAndCodeModels { get; set; }
     }
 }
